Move get_flux_points
